agregado de favoritos

+++ src/main/java/com/example/demo/entity/Review.java
@@ -24,7 +24,7 @@ public class Review {
     private LocalDateTime date;
 
     @Column
-    private double value;
+    private int value;
 
     @Column
     private String comment;
